添加一个强制转义，指定pow函数参数类型。

+++ purenessscopeserver/purenessscopeserver/PurenessScopeServer/Common/MessageBlockManager.h
@@ -48,7 +48,7 @@ public:
     {
         for(uint32 i = 0; i < MAX_MEMORY_BLOCK_COUNT; i++)
         {
-            m_MemoryBlock_List[i].m_u4ListSize = (uint32)pow((double)2, i);;
+            m_MemoryBlock_List[i].m_u4ListSize = (uint32)pow((double)2, (int32)i);;
         }
     }
